+++ common/src/main/java/al/alec/custommachinerybotania/integration/kubejs/ManaRequirementJS.java
@@ -6,15 +6,15 @@ import fr.frinn.custommachinery.api.requirement.RequirementIOMode;
 
 public interface ManaRequirementJS extends RecipeJSBuilder {
 
-  default RecipeJSBuilder requireMana(long mana) {
+  default RecipeJSBuilder requireMana(int mana) {
     return addManaRequirement(RequirementIOMode.INPUT, mana);
   }
 
-  default RecipeJSBuilder produceMana(long mana) {
+  default RecipeJSBuilder produceMana(int mana) {
     return addManaRequirement(RequirementIOMode.OUTPUT, mana);
   }
 
-  default RecipeJSBuilder addManaRequirement(RequirementIOMode mode, long mana) {
+  default RecipeJSBuilder addManaRequirement(RequirementIOMode mode, int mana) {
     if (mana < 0)
       return error("Mana value cannot be negative");
     return addRequirement(new ManaRequirement(mode, mana));
