+++ common/src/main/java/al/alec/custommachinerybotania/integration/kubejs/ManaRequirementPerTickJS.java
@@ -6,15 +6,15 @@ import fr.frinn.custommachinery.api.requirement.RequirementIOMode;
 
 public interface ManaRequirementPerTickJS extends RecipeJSBuilder {
 
-  default RecipeJSBuilder requireManaPerTick(long mana) {
+  default RecipeJSBuilder requireManaPerTick(int mana) {
     return addManaPerTickRequirement(RequirementIOMode.INPUT, mana);
   }
 
-  default RecipeJSBuilder produceManaPerTick(long mana) {
+  default RecipeJSBuilder produceManaPerTick(int mana) {
     return addManaPerTickRequirement(RequirementIOMode.OUTPUT, mana);
   }
 
-  default RecipeJSBuilder addManaPerTickRequirement(RequirementIOMode mode, long mana) {
+  default RecipeJSBuilder addManaPerTickRequirement(RequirementIOMode mode, int mana) {
     if (mana < 0)
       return error("Mana value cannot be negative");
     return addRequirement(new ManaRequirementPerTick(mode, mana));
